Add color theme feature.
This allows you to choose between preset themes stored in .ini files
in the layout.  While I was at it, I changed other things in the
Settings panel to auto-populate from the layout rather than have
hard-coded versions.  So there's no longer a need to recompile
Quackle to add new languages or alphabets to get it to show up in
the popup menu.