Implement FileOpen event
This allows Quackle to respond to system file open
events.  So, if .gcg is associated with Quackle, then
double-clicking a .gcg file will now make it open in
Quackle (whether it's running or not).  And drag and
drop, etc.

This definitely fixes issues trying to open files with
double-click on Mac.  Windows doesn't have the
.gcg association set up yet, so code is untested
there, but it should be platform independent.