Add versioning of DAWGs.
If we're going to start writing these into user directories,
then we'd better start versioning them so we don't end
up generating bugs in the future.

LexiconParameters::loadDawg() implements a tiny
class factory which allows backward compatibility
of DAWGs.  I'll soon be adding a "version 1" in
addition to the legacy "version 0".  For now, version 1
is just dummied in.