Set up defaults strategies.
So, basically, the dictionary is queried to see if it's
"English-like".  Which is to say that it uses the same
alphabet.  If so, then it will fall back to use twl06
strategies, minus the superleaves (the so-called
"default_english").  Otherwise, it uses "default" strategies,
which right now is simply a worths file that sets the
blank to be 30.

It's not great.  But better than before.