Add a scoring option preference.
A checkbox in the preferences dialog now allows
you to configure Quackle so that plays with illegal
words which are left unchallenged score zero.

Obviously, off by default.  But useful for entering
games as part of the Marty Gabriel/Scott Garner
world record scoring attempt.  This is apparently
what the Guinness folks are looking for.