Fix problems with "ex n" where n is a number.
Found a crash where you enter "ex 4" twice,
and it would crash.  This led me to look up
how this was implemented, and it wasn't
very robust.  It also didn't save properly in
the GCG.  So I created a new move type
which I called a BlindExchange and implemented
it throughout the system.