Fix problems entering games with unknown racks.
If you're entering games with unknown racks, Quackle can
get confused about which player has how many tiles in the
end game.  This is because the racks may have been set to
have less than 7 tiles, and some of the tiles "in the bag" may
actually belong on another player's rack.  Fixed this by...

* Adding a mechanism which tracks the actual count of
tiles in the bag and on the rack in GamePosition.  This
count is independent of what's actually on the rack.
* If the game is about to end because the bag and rack
are empty, check to see if the rack *should* have been
empty.  If not, pull tiles from another player's rack.
This code might fail for games of more than two players.
* If the game should be ending because the bag and
rack are empty, but the rack isn't actually empty, then
dump the files to another player's rack.  Once again,
this code doesn't take into account >2 players.