Fixes to allow bad endgames to be loaded from GCGs.
Not perfect, but it now does a decent job of not corrupting GCGs when
it loads one where players have messed up tile counts/drawing in the
end-game.  It also tries to annotate these for the player and in any
GCG file that's round-tripped.

Also, a lot of C++11-izing of ranged iterators.