Move simulate code to a new static function.
The code running the simulation is now
disconnected from the Simulation class,
with appropriate members being copied into
a SimmedMoveConstants field (read-only,
thread-safe) and a SimmedMoveMessage
(the class for communicating per-sim data
to and from external threads).

The newly added static function,
Simulate::simulateOnePosition(), should now
be thread-safe and usable as a run function
for a thread.