Sims are now threaded.
Defaults to two threads right now.  This
can be changed in the Simulator constructor
for the time being.  An interface will be
added for this soon.

For an example I tried, I got the following times on 500 simulations in a
macOS release build:

* Quackle 1.0.3 - 26 seconds
* master branch, 1 thread - 27 seconds
* master branch, 2 threads - 19 seconds
* master branch, 4 threads - 15 seconds

This isn't the most efficient use of threads.
It distributes all of the plays being simmed
into a thread pool, but only for one simulation.
Then it comes up for air and tries again.
This was the easiest to implement robustly
given the current architecture.