Clean up duplicate strategy files.
There's only two distinct superleaves files, but there were
more copies in the layout and it's a huge file.  Now, lexica
beginning with the text "csw" will fall back to the strategy files
in the "csw" directory before going to default_english.  And so,
the Collins superleaves is in strategy/csw and other dictionaries
default directly to strategy/default_english.

Also clean up dupes of other files which didn't take much space,
but at least it clarifies what strategy files are unique.

Minor change to the C++ code to detect the "csw" directory.