Hard-code insertion sort in String::alphabetize().
std::sort has to be fast for both large and small collections, while we
know our strings to be sorted are very small (<= 7 letters) and already
mostly-sorted, which is the ideal case for insertion sort.

Thus, we can hard-code an insertion sort instead, avoiding the checks
that std::sort does for merge sort, calls to memcpy() etc. On a very
rough profile, this speeds up String::alphabetize() by about 85%
(GCC 10.1, Kaby Lake).