Make vcvalues static.
When the vcvalues array is not static but merely const, the compiler
(at least GCC) will spend time copying it all onto the stack before
deferencing it. This is not a huge loss, especially since English
doesn't hit this code path at all, but it's still unneeded and shows
up in the profiles for other languages.