Skip deduplication of exchanges when not needed.
Generator::exchange() uses an std::map to keep track of equivalent exchange
moves we've already considered, so as to not create duplicates in the move list
(e.g. if we have AABC on the rack and drop one of the As, it doesn't matter
whether it's the first or the second A; the moves are equivalent).
However, std::map is node-based in order to guarantee pointer stability,
and thus calls malloc/free for each insertion. This is slow, and even more so
on platforms with slow mallocs (e.g. macOS). Furthermore, when using many search
threads, concurrent malloc can become a choke point.

In most cases, we don't really care about recording the moves; we just
want the best one. Thus, don't bother deduplicating unless we actually
record, which speeds up Generator::exchange() to roughly five times the
speed (GCC 10.1, Linux -- using the standard glibc allocator, which is
fairly fast, so might be more on other platforms).

Also change from std::map<LetterString, bool> to std::set<LetterString>,
as the value was always true anyway.