Streamline Rack::unload().
There's no need to generate a new string to hold the new contents
of the rack; instead, we can move tiles to the left to fill in
the blank spots. This makes for less copying, and also significantly
fewer branch mispredictions. Speeds up the function by roughly 50%
(GCC 10.1, Kaby Lake).