Spend less time in counting used tiles.
Two related optimizations to String::usedTiles():

First, replace the branches, which are poorly-predicted, with a lookup table.
This makes for faster code with no branches, which speeds up the function by
roughly 25%.

Second, add a new function numUsedTiles() using the same table, that doesn't
need to construct an intermediate string (which makes for a call to memcpy).
In particular, CatchallEvaluator (used for non-English languages, I believe?)
wants to know how many tiles are left in the bag, and calls
String::usedTiles().length() as part of that. Making a specialized
(also branch-free) version that only counts instead of constructing any strings
speeds up generation _overall_ by 6–7% in these cases (GCC 10.1, Kaby Lake).

It might be possible to optimize Rack::operator-(Move) to also not construct any
intermediate strings, by fusing this logic into Rack::unload(). It is the main
remaining caller of usedTiles().