Treat GADDAG nodes as a 32-bit unit.
Each GADDAG node, as stored on disk, is essentially a 32-bit bit field,
and (after byte-swapping), can be treated as such. Instead of writing
code to manually piece together the bytes when we need the upper 24 bits,
we can simply treat it as a uint32_t, which saves a few repeated loads
and some shifting.

Speeds up move generation by about 4% in a local benchmark (GCC 10.1,
Kaby Lake).