Make SuperLeavesMap a hash table.
The map converting (super)leaves to value is in several million entries,
which means that std::map needs a lot of FixedString compares, which seem
to be rather expensive (more than 30% of total CPU time used).

Simply convert into using std::unordered_map. No current compiler has a
very efficient implementation of these hash tables (there are independent
implementations that are much faster, some with looser design constraints that
would fit Quackle well), yet the speed of ScorePlusLeaveEvaluator::leaveValue
with superleaves easily triples by this simple change, for an overall 50–55%
speedup on TWL06. For hash function, we simply reuse std::string's. If we are
guaranteed C++17, std::string_view's would probably be a better fit, but due to
the short string optimizations, we don't actually create anything on the heap.

Note that in the process, we need to move several operator== into the
Quackle namespace. Due to the mechanics of overload resolution and ADL,
all free-standing operator overloads need to be in the namespace of one of its
arguments, but Quackle would put e.g. operator==(Quackle::Move a,
Quackle::Move b) into the global namespace. This worked fine as long as all
calling code was either in the global namespace or from Quackle::, but now,
there will be calls from within std::unordered_map (ie., in std::, where
there are other operator== candidates, which makes the namespace search
stop before it considers candidates in the global namespace -- even though
none of them match), and it needed to be fixed.