Deduplicate tails in GADDAGs.
The GADDAG as currently implemented is a pure trie. This means that there
are lots of nodes with identical contents; e.g., in TWL09 (4.2M nodes),
there are 388,839 nodes that all signify that something can end in -s.
We can deduplicate these and point every parent to the same node, removing
a lot of them. This also works for more complex nodes, e.g. 32,678 nodes
specify something that can end in -d or -s; they can also be merged into
one.

Once we have deduplicated such simple tails, we can deduplicate even more;
e.g., when we have joined 99,586 -g tails, we can join 68,870 -ng tails,
and from there we can join 10,413 -ed/-es/-ing tails (assuming we've also
deduplicated -ed and -es, of course). We continue running such deduplication
passes until there's nothing more to be done and the graph is minimal.
All in all, a typical word list gets a full 80% of nodes removed in this
fashion, with correspondingly reduced RAM and disk requirements.

In itself, this might not be that interesting, since RAM and disk are both
cheap these days. However, this also means that the cache footprint of the
GADDAG goes down by the same amount; typical L3 cache in an x86 CPU is
8 MB, and reducing a GADDAG from 16 to 4 MB means vastly fewer L3 misses
in the word generator (and also somewhat fewer L1 and L2 misses).
This speeds up the AI by about 2%; I was maybe surprised it wasn't more than
that, but it seems the working set is already fairly constrained by what's
already on the board, so L1 and L2 misses might be a worse issue than L3
misses, and we're still far away from fitting into L2. (Also, of course,
other things like superleaves can take up a lot of space in the cache.)

It is possible to pack the most popular nodes together so that they are
not spread over a bunch of cachelines shared with uninteresting stuff,
but that doesn't seem to actually help much in practice; the hardware
cache miss counters look better, but actual self-play is not appreciably
faster. If one wants to go down this route, it's possible that some kind
of tricky heuristics are needed in order to not disturb the locality in
the root of the tree and/or the less-popular leaves.

We always deduplicate towards the entry with the highest index, to avoid
getting negative offsets. One could wonder if this does not reduce the
benefits of the previous patch (which reduces offsets to allow bigger word
lists as GADDAGs), as deduplication means that the offsets will get larger.
This is true, but the effect is more than offset by the sheer node count
reductions; e.g. OSPS, which had the biggest problem before, now loses 93%
of its nodes (down to 4M nodes), which means no matter what ordering we have,
we will fit comfortably within the 16M offset limit. If this proves to be a
problem for even larger word lists in the future, it would be possible to
limit deduplication to the cases where it doesn't create new offsets,
or somehow shuffle the nodes more carefully around.

Deduplicating tails like this in a GADDAG is not novel; the original GADDAG
paper also talks about similar minimization. It also has an algorithm for doing
partial deduplication right as a word is inserted, in order to reduce the amount
of time spent in the full deduplication algorithm. After all, most of these
duplicates stem from the fact that GADDAG inserts the same word many times;
e.g. just for HELPS, the -s node is inserted five times with various prefixes,
and those multi-inserts are easy to deduplicate. However, we don't implement said
algorithm, which is just an optimization to reduce the GADDAG build time;
a large word list may take a minute or two to deduplicate, but it's a one-time
process.

We don't do partial deduplication (e.g., it would be possible to point into
the middle of another node, so that the -s node could simply be a pointer
midway into the -d/-s node) since the gains don't appear to justify the
complexity, but it could be added later.

DAWGs should also be given the same treatment (they are also pure tries
in Quackle's implementation); this has also not been done.