Add cmake/GitHub Actions-based CI, dump Travis-CI.
Travis has been throwing a lot of false failures on the Mac build, which was
always super-slow anyway.  And now it's been suggested that Travis ain't gonna
last in the open-source-free-for-use CI business.  So out it goes.

In its place, GitHub Actions.  Just one for now, which runs the build on
pushes to master and on PRs (I hope...haven't tested the PR bit yet).
Notes about the new system...

* It uses the cmake build.  Next step...document the cmake system and
  de-document the qmake system.
* It tests against both Qt 5.12.0 and 5.15.1.
* Mac, Windows (x64, x86) and Linux (Ubuntu 18.04, 20.04)
* On Linux, it also builds clang and g++.
* In all of the above cases, it's just pretty much default compilers.  Which
  is probably good...particularly on Linux, where it's not so common for
  non-developers to be running major compiler updates.
* All cmake builds use ninja and cache their Qt downloads, so the result is
  fairly fast.
* No builds of makegaddag, gaddagize, makeminidawg, iotest, or encodeleaves
  yet, as I haven't migrated those to cmake.  Need to get on that.