Fix up std:: namespace hygiene.
* Remove all uses of `using namespace std` from header files (although a
  few specific cases of things like `using std::vector` remain)
* Move all existing uses of `using namespace std` in cpp files to after the
  header includes
* Make coordinating changes so it builds again.

Qt6 builds on Windows were breaking because of namespace collision between
std::byte and some version of `byte` in one of Qt's namespaces.  That breakage
is now fixed.

It would be good to systematically clean up all `using` declarations in header
files to ensure nothing is leaking out, but that's for another time.